:recycle: Separate game and app stores (meh)
