+++ src/store/game-state.ts
@@ -1,37 +1,20 @@
-import { writable, derived, get } from 'svelte/store'
-import type { Writable, Updater, Readable } from 'svelte/store'
-import { writable as storageWritable } from 'svelte-local-storage-store'
-import type { Board, GameMode, Stats } from '$lib/data-model'
+import type { Readable, Updater, Writable } from 'svelte/store'
+import { derived, get, writable } from 'svelte/store'
+import type { Board, GameMode } from '$lib/data-model'
 import {
 	createNewBoard,
 	getValidLetterBounds,
 	getValidLetters,
-	newStats,
 	ROWS,
 	scoreTile,
 } from '$lib/data-model'
-
-export const storeVersion: Writable<number> = storageWritable('wp-version', 0)
-export const storedLocale: Writable<string> = storageWritable('wp-locale', '')
-
-export const openScreen: Writable<null | 'options' | 'tutorial' | 'results'> = writable(null)
+import { writable as storageWritable } from 'svelte-local-storage-store'
 
 export const answerDaily: Writable<string> = storageWritable('wp-answer', '')
 export const answerRandom: Writable<string> = storageWritable('wp-answerRandom', '')
 
 export const guessesDaily: Writable<string[]> = storageWritable('wp-guesses', [])
 export const guessesRandom: Writable<string[]> = storageWritable('wp-guessesRandom', [])
-
-// TODO: Consolidate options to single object?
-export const highContrast: Writable<boolean> = storageWritable('wp-highContrast', false)
-export const showAllHints: Writable<boolean> = storageWritable('wp-showAllHints', false)
-const hardModeStored: Writable<boolean> = storageWritable('wp-hardMode', false)
-export const hardMode: Readable<boolean> = derived(
-	hardModeStored,
-	($hardModeStored) => $hardModeStored
-)
-export const swapEnterBackspace: Writable<boolean> = storageWritable('wp-swapEnterBackspace', false)
-export const keyboardLayout: Writable<string> = storageWritable('wp-keyboardLayout', 'qwerty')
 
 export const lastPlayedDailyWasHard: Writable<boolean> = storageWritable(
 	'wp-lastPlayedWasHard',
@@ -42,14 +25,15 @@ export const lastPlayedRandomWasHard: Writable<boolean> = storageWritable(
 	false
 )
 
+const hardModeStored: Writable<boolean> = storageWritable('wp-hardMode', false)
+export const hardMode: Readable<boolean> = derived(
+	hardModeStored,
+	($hardModeStored) => $hardModeStored
+)
 export const changeHardMode = (changeTo: boolean) => {
 	if (!get(gameFinished) && get(guesses).length > 0) throw "Can't change that during a game!"
 	hardModeStored.set(changeTo)
 }
-
-export const lastPlayedDaily: Writable<number> = storageWritable('wp-lastPlayedDaily', -1)
-
-export const stats: Writable<Stats> = storageWritable('wp-stats', newStats())
 
 export const gameMode: Writable<GameMode> = writable('daily')
